ref(rust): Split factory into separate module
consumer.rs is getting unwieldy, split the factory into a separate file.
Also passes the whole StorageConfig object to the factory - we will
need more data from this object later to produce to the commit log.
